requirements.txt
# Dependencias principales del proyecto
boto3>=1.34.0
click>=8.1.7
python-dotenv>=1.0.0
pydantic>=2.5.0

# Procesamiento de documentos
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
pandas>=2.1.0
openpyxl>=3.1.0
Pillow>=10.1.0

# Utilidades
rich>=13.7.0
requests>=2.31.0
